requirements.txt

streamlit
google-genai
yfinance
pandas_ta
duckduckgo-search